research/kg_hyp_emb/requirements.txt
numpy>=1.17
tensorflow_gpu>=2.0.0
